+++ framy_generator/lib/generator/app_bar_generator.dart
@@ -1,10 +1,15 @@
 String generateAppBar() => '''
 class FramyAppBar extends StatelessWidget with PreferredSizeWidget {
+  final Widget leading;
+
+  const FramyAppBar({Key key, @required this.leading}) : super(key: key);
+  
   @override
   Widget build(BuildContext context) {
     return AppBar(
       key: Key('FramyAppBar'),
       title: Text('Framy App'),
+      leading: leading,
       actions: [
         IconButton(
           key: ValueKey('FramyAppBarSettingsButton'),