+++ test_apps/weight_tracker/lib/main.app.framy.dart
@@ -50,15 +50,16 @@ class _FramyAppState extends State<FramyApp> {
   bool _wrapWithScaffold = true;
   bool _wrapWithCenter = false;
   bool _wrapWithSafeArea = false;
+  bool _showNavigationMenu = true;
 
-  void set wrapWithScaffold(bool value) =>
-      setState(() => _wrapWithScaffold = value);
+  set wrapWithScaffold(bool value) => setState(() => _wrapWithScaffold = value);
 
-  void set wrapWithCenter(bool value) =>
-      setState(() => _wrapWithCenter = value);
+  set wrapWithCenter(bool value) => setState(() => _wrapWithCenter = value);
 
-  void set wrapWithSafeArea(bool value) =>
-      setState(() => _wrapWithSafeArea = value);
+  set wrapWithSafeArea(bool value) => setState(() => _wrapWithSafeArea = value);
+
+  void toggleNavigationMenu() =>
+      setState(() => _showNavigationMenu = !_showNavigationMenu);
 
   @override
   Widget build(BuildContext context) {
@@ -67,6 +68,7 @@ class _FramyAppState extends State<FramyApp> {
         wrapWithScaffold: _wrapWithScaffold,
         wrapWithCenter: _wrapWithCenter,
         wrapWithSafeArea: _wrapWithSafeArea,
+        showNavigationMenu: _showNavigationMenu,
         child: MaterialApp(
           key: Key('FramyApp'),
           debugShowCheckedModeBanner: false,
@@ -82,6 +84,7 @@ class FramyAppSettings extends InheritedWidget {
   final bool wrapWithScaffold;
   final bool wrapWithCenter;
   final bool wrapWithSafeArea;
+  final bool showNavigationMenu;
 
   const FramyAppSettings({
     Key key,
@@ -89,6 +92,7 @@ class FramyAppSettings extends InheritedWidget {
     @required this.wrapWithScaffold,
     @required this.wrapWithCenter,
     @required this.wrapWithSafeArea,
+    @required this.showNavigationMenu,
   })  : assert(child != null),
         super(key: key, child: child);
 
@@ -100,7 +104,8 @@ class FramyAppSettings extends InheritedWidget {
   bool updateShouldNotify(FramyAppSettings old) =>
       old.wrapWithScaffold != wrapWithScaffold ||
       old.wrapWithCenter != wrapWithCenter ||
-      old.wrapWithSafeArea != wrapWithSafeArea;
+      old.wrapWithSafeArea != wrapWithSafeArea ||
+      old.showNavigationMenu != showNavigationMenu;
 }
 
 Route onGenerateRoute(RouteSettings settings) {
@@ -139,12 +144,24 @@ class FramyLayoutTemplate extends StatelessWidget {
     return LayoutBuilder(
       builder: (context, constraints) {
         final isSmallDevice = constraints.maxWidth < 1000;
+        final leading = isSmallDevice
+            ? null
+            : IconButton(
+                icon: const Icon(Icons.menu),
+                onPressed: () =>
+                    framyAppStateKey.currentState.toggleNavigationMenu(),
+                tooltip: FramyAppSettings.of(context).showNavigationMenu
+                    ? 'Close navigation menu'
+                    : 'Open navigation menu',
+              );
         return Scaffold(
-          appBar: FramyAppBar(),
+          appBar: FramyAppBar(leading: leading),
           body: Row(
             crossAxisAlignment: CrossAxisAlignment.start,
             children: [
-              if (!isSmallDevice) FramyDrawer(),
+              if (!isSmallDevice &&
+                  FramyAppSettings.of(context).showNavigationMenu)
+                FramyDrawer(),
               Expanded(
                 child: child,
               ),
@@ -158,11 +175,16 @@ class FramyLayoutTemplate extends StatelessWidget {
 }
 
 class FramyAppBar extends StatelessWidget with PreferredSizeWidget {
+  final Widget leading;
+
+  const FramyAppBar({Key key, @required this.leading}) : super(key: key);
+
   @override
   Widget build(BuildContext context) {
     return AppBar(
       key: Key('FramyAppBar'),
       title: Text('Framy App'),
+      leading: leading,
       actions: [
         IconButton(
           key: ValueKey('FramyAppBarSettingsButton'),