+++ framy_generator/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [0.1.10]
 
 * Changed showing settings dialog to work for showcase app #116
+* Added hiding navigation drawer on big screens #113
 
 ## [0.1.9]